v-1

+++ src/Helper/blog.js
@@ -1,4 +1,5 @@
 import db from "../config/db";
+import {isAuthenticated} from "./user";
 
 export const getAllBlogFromDB = () => {
 	return fetch(`${db}/all/blog`)
@@ -7,17 +8,25 @@ export const getAllBlogFromDB = () => {
 };
 
 export const getBlogInfoByBlogId = (blogId) => {
-	return fetch(`${db}/blog/${blogId}`)
+	return fetch(`${db}/blog/${blogId}`, {
+		method: "GET",
+	})
 		.then((result) => result.json())
 		.catch((error) => console.error(error));
 };
 
 export const addBlogtoDB = (input, userId) => {
+	const {token} = isAuthenticated();
+	if (!token) {
+		alert("Unauthorized to access it");
+		return;
+	}
 	return fetch(`${db}/add/blog/${userId}`, {
 		method: "POST",
 		headers: {
 			Accept: "application/json",
 			"Content-Type": "application/json",
+			Authorization: `Bearer ${token}`,
 		},
 		body: JSON.stringify(input),
 	})
@@ -42,11 +51,17 @@ export const getBlogInfoFromDb = (blogId) => {
 };
 
 export const updateBlogFromFB = (input, userId, blogId) => {
+	const {token} = isAuthenticated();
+	if (!token) {
+		alert("Unauthorized to access it");
+		return;
+	}
 	return fetch(`${db}/update/comment/${userId}/${blogId}`, {
 		method: "PUT",
 		headers: {
 			Accept: "application/json",
 			"Content-Type": "application/json",
+			Authorization: `Bearer ${token}`,
 		},
 		body: JSON.stringify(input),
 	})
@@ -55,11 +70,17 @@ export const updateBlogFromFB = (input, userId, blogId) => {
 };
 
 export const likeBlogByBlogIdDB = (userId, blogId) => {
+	const {token} = isAuthenticated();
+	if (!token) {
+		alert("Unauthorized to access it");
+		return;
+	}
 	return fetch(`${db}/add/like/${userId}/${blogId}`, {
 		method: "GET",
 		headers: {
 			Accept: "application/json",
 			"Content-Type": "application/json",
+			Authorization: `Bearer ${token}`,
 		},
 	})
 		.then((result) => result.json())
@@ -67,11 +88,17 @@ export const likeBlogByBlogIdDB = (userId, blogId) => {
 };
 
 export const removelikeBlogByBlogIdDB = (userId, blogId) => {
+	const {token} = isAuthenticated();
+	if (!token) {
+		alert("Unauthorized to access it");
+		return;
+	}
 	return fetch(`${db}/remove/like/${userId}/${blogId}`, {
 		method: "GET",
 		headers: {
 			Accept: "application/json",
 			"Content-Type": "application/json",
+			Authorization: `Bearer ${token}`,
 		},
 	})
 		.then((result) => result.json())
@@ -79,11 +106,17 @@ export const removelikeBlogByBlogIdDB = (userId, blogId) => {
 };
 
 export const dislikeBlogByBlogIdDB = (userId, blogId) => {
+	const {token} = isAuthenticated();
+	if (!token) {
+		alert("Unauthorized to access it");
+		return;
+	}
 	return fetch(`${db}/add/dislike/${userId}/${blogId}`, {
 		method: "GET",
 		headers: {
 			Accept: "application/json",
 			"Content-Type": "application/json",
+			Authorization: `Bearer ${token}`,
 		},
 	})
 		.then((result) => result.json())
@@ -91,11 +124,35 @@ export const dislikeBlogByBlogIdDB = (userId, blogId) => {
 };
 
 export const removedislikeBlogByBlogIdDB = (userId, blogId) => {
+	const {token} = isAuthenticated();
+	if (!token) {
+		alert("Unauthorized to access it");
+		return;
+	}
 	return fetch(`${db}/remove/dislike/${userId}/${blogId}`, {
 		method: "GET",
 		headers: {
 			Accept: "application/json",
 			"Content-Type": "application/json",
+			Authorization: `Bearer ${token}`,
+		},
+	})
+		.then((result) => result.json())
+		.catch((error) => console.error(error));
+};
+
+export const deleteBlogByBlogId = (blogId) => {
+	const {token} = isAuthenticated();
+	if (!token) {
+		alert("Unauthorized to access it");
+		return;
+	}
+	return fetch(`${db}/blog/${blogId}`, {
+		method: "DELETE",
+		headers: {
+			Accept: "application/json",
+			"Content-Type": "application/json",
+			Authorization: `Bearer ${token}`,
 		},
 	})
 		.then((result) => result.json())
@@ -105,11 +162,17 @@ export const removedislikeBlogByBlogIdDB = (userId, blogId) => {
 //Bookmark
 
 export const addBookMarkDB = (userId, blogId) => {
+	const {token} = isAuthenticated();
+	if (!token) {
+		alert("Unauthorized to access it");
+		return;
+	}
 	return fetch(`${db}/add/bookmark/${userId}/${blogId}`, {
 		method: "GET",
 		headers: {
 			Accept: "application/json",
 			"Content-Type": "application/json",
+			Authorization: `Bearer ${token}`,
 		},
 	})
 		.then((result) => result.json())
@@ -117,11 +180,17 @@ export const addBookMarkDB = (userId, blogId) => {
 };
 
 export const RemoveBookMarkDB = (userId, blogId) => {
+	const {token} = isAuthenticated();
+	if (!token) {
+		alert("Unauthorized to access it");
+		return;
+	}
 	return fetch(`${db}/remove/bookmark/${userId}/${blogId}`, {
 		method: "DELETE",
 		headers: {
 			Accept: "application/json",
 			"Content-Type": "application/json",
+			Authorization: `Bearer ${token}`,
 		},
 	})
 		.then((result) => result.json())
@@ -129,24 +198,36 @@ export const RemoveBookMarkDB = (userId, blogId) => {
 };
 
 export const allBookMarkByBlogIdDB = (blogId) => {
+	const {token} = isAuthenticated();
+	if (!token) {
+		alert("Unauthorized to access it");
+		return;
+	}
 	return fetch(`${db}/all/bookmark/blog/${blogId}`, {
 		method: "GET",
 		headers: {
 			Accept: "application/json",
 			"Content-Type": "application/json",
 		},
+		Authorization: `Bearer ${token}`,
 	})
 		.then((result) => result.json())
 		.catch((error) => console.error(error));
 };
 
 export const allBookMarkByUserIdDB = (userId) => {
+	const {token} = isAuthenticated();
+	if (!token) {
+		alert("Unauthorized to access it");
+		return;
+	}
 	return fetch(`${db}/all/bookmark/user/${userId}`, {
 		method: "GET",
 		headers: {
 			Accept: "application/json",
 			"Content-Type": "application/json",
 		},
+		Authorization: `Bearer ${token}`,
 	})
 		.then((result) => result.json())
 		.catch((error) => console.error(error));
